python/projects/py-invest/pyproject.toml
[project]
name = "invest-ai"
version = "0.1.0"
description = "LLM-powered stock analysis report platform"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]

dependencies = [
    # Web framework
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",

    # LLM orchestration
    "langchain>=0.1.0",
    "langchain-openai>=0.0.5",
    "langgraph>=0.0.20",

    # Data collection
    "yfinance>=0.2.36",
    "akshare>=1.12.0",
    "pandas>=2.1.0",
    "numpy>=1.26.0",
    "requests>=2.31.0",
    "aiohttp>=3.9.0",

    # Technical analysis
    "pandas-ta>=0.3.14b",
    "ta-lib>=0.4.28",

    # Data storage
    "sqlalchemy>=2.0.25",
    "asyncpg>=0.29.0",
    "redis>=5.0.0",

    # Utilities
    "python-dotenv>=1.0.0",
    "structlog>=24.1.0",
    "tenacity>=8.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
    "httpx>=0.26.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
py-invest = "cli:main"

[tool.hatch.build.targets.wheel]
packages = ["apps", "modules", "agents", "core", "storage"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "C4"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
